+++ src/backend/common/half.hpp
@@ -47,6 +47,7 @@ using uint16_t = unsigned short;
 #include <type_traits>
 
 #include <limits>
+
 #endif
 
 namespace common {
