+++ src/backend/opencl/types.cpp
@@ -10,6 +10,7 @@
 #include <types.hpp>
 
 #include <common/half.hpp>
+#include <common/util.hpp>
 #include <type_util.hpp>
 
 #include <cmath>
@@ -17,35 +18,39 @@
 #include <string>
 
 using common::half;
+using common::toString;
+
+using std::isinf;
+using std::stringstream;
 
 namespace opencl {
 
 template<typename T>
 inline std::string ToNumStr<T>::operator()(T val) {
     ToNum<T> toNum;
-    return std::to_string(toNum(val));
+    return toString(toNum(val));
 }
 
 template<>
 std::string ToNumStr<float>::operator()(float val) {
     static const char *PINF = "+INFINITY";
     static const char *NINF = "-INFINITY";
-    if (std::isinf(val)) { return val < 0.f ? NINF : PINF; }
-    return std::to_string(val);
+    if (isinf(val)) { return val < 0.f ? NINF : PINF; }
+    return toString(val);
 }
 
 template<>
 std::string ToNumStr<double>::operator()(double val) {
     static const char *PINF = "+INFINITY";
     static const char *NINF = "-INFINITY";
-    if (std::isinf(val)) { return val < 0. ? NINF : PINF; }
-    return std::to_string(val);
+    if (isinf(val)) { return val < 0. ? NINF : PINF; }
+    return toString(val);
 }
 
 template<>
 std::string ToNumStr<cfloat>::operator()(cfloat val) {
     ToNumStr<float> realStr;
-    std::stringstream s;
+    stringstream s;
     s << "{" << realStr(val.s[0]) << "," << realStr(val.s[1]) << "}";
     return s.str();
 }
@@ -53,7 +58,7 @@ std::string ToNumStr<cfloat>::operator()(cfloat val) {
 template<>
 std::string ToNumStr<cdouble>::operator()(cdouble val) {
     ToNumStr<double> realStr;
-    std::stringstream s;
+    stringstream s;
     s << "{" << realStr(val.s[0]) << "," << realStr(val.s[1]) << "}";
     return s.str();
 }
@@ -64,8 +69,8 @@ std::string ToNumStr<half>::operator()(half val) {
     using namespace common;
     static const char *PINF = "+INFINITY";
     static const char *NINF = "-INFINITY";
-    if (common::isinf(val)) { return val < 0.f ? NINF : PINF; }
-    return common::to_string(val);
+    if (isinf(val)) { return val < 0.f ? NINF : PINF; }
+    return toString(val);
 }
 
 template<>
@@ -73,8 +78,8 @@ template<>
 std::string ToNumStr<half>::operator()<float>(float val) {
     static const char *PINF = "+INFINITY";
     static const char *NINF = "-INFINITY";
-    if (common::isinf(half(val))) { return val < 0.f ? NINF : PINF; }
-    return std::to_string(val);
+    if (isinf(half(val))) { return val < 0.f ? NINF : PINF; }
+    return toString(val);
 }
 
 #define INSTANTIATE(TYPE) template struct ToNumStr<TYPE>
